Visual changes to log button
